pyproject.toml
[tool.poetry]
name = "computer-vision-benchmarks"
version = "0.1.0"
description = ""
authors = ["Tom Pope"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tensorflow-macos = "^2.16.2"
tensorflow-metal = "^1.1.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
